backend/requirements.txt
fastapi
uvicorn
wikipedia-api
nltk
faiss-cpu
numpy
sentence-transformers
torch
python-dotenv
